chore: decomposited footer to components
