+++ src/components/common/Footer/Footer.module.scss
@@ -17,101 +17,12 @@
   align-items: center;
 }
 
-.logo {
-  color: $white-text;
-}
-.logo img:hover {
-  content: url('/svg/logo-hover.svg');
-}
-/////////////////
-.footerMenu ul {
-  list-style: none;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.footerMenu li {
-  display: inline-block;
-}
-
-.footerMenu a {
-  color: $white-text;
-}
-
-.footerMenu a:hover {
-  color: $background-two;
-}
-////////////////// //
-.footerLinks ul {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.footerLinks a {
-  color: $white-text;
-  position: relative;
-}
-
-.footerLinks a:hover {
-  color: $background-two;
-}
-
-.footerLinks a::after {
-  content: '';
-  position: absolute;
-  left: 0;
-  bottom: -5px;
-  width: 100%;
-  height: 1px;
-  background-color: $white-text;
-}
-.footerLinks a:hover::after {
-  background-color: $white-text;
-}
-///////////////
 .footerBottom {
   display: flex;
   justify-content: space-between;
   align-items: center;
 }
 
-.footerContacts ul {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.footerContacts a {
-  color: $white-text;
-  white-space: nowrap;
-  display: flex;
-  align-items: center;
-}
-
-.footerContacts img {
-  margin-right: 10px;
-}
-
-///////////
-.footerSocial {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 2.5rem;
-  margin-top: 1rem;
-  width: 14rem;
-}
-
-.footerSocial a {
-  color: $white-text;
-}
-.footerSocial img {
-  margin: 0;
-}
-
-///////////////
 .footerCopyright {
   margin-top: 3rem;
   color: $white-text;
@@ -171,24 +82,7 @@
     flex-direction: column;
     align-items: flex-start;
   }
-  .logo {
-    margin: 0 auto;
-    margin-bottom: 1.5rem;
-  }
-  .footerMenu {
-    margin-bottom: 1.5rem;
-  }
-  .footerLinks {
-    margin-bottom: 1.5rem;
-  }
 
-  // має бути по середині
-  .footerSocial {
-    width: 100%;
-    position: relative;
-    left: 4.8rem;
-  }
-  ////
   .footerCopyright p {
     font-size: 0.875rem;
     font-weight: 500;
