+++ src/components/common/Footer/components/FooterContacts/FooterContacts.module.scss
@@ -1,0 +1,39 @@
+@import '@/theme/variables';
+@import '@/theme/mixins';
+
+.wrapper {
+  ul {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+  }
+  a {
+    color: $white-text;
+    white-space: nowrap;
+    display: flex;
+    align-items: center;
+  }
+  img {
+    margin-right: 10px;
+  }
+}
+
+.social {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 2.5rem;
+  margin-top: 1rem;
+  width: 14rem;
+  @media (max-width: $mobile) {
+    width: 100%;
+    position: relative;
+    left: 4.8rem;
+  }
+  a {
+    color: $white-text;
+  }
+  img {
+    margin: 0;
+  }
+}
