+++ src/components/common/Footer/components/FooterDocuments/FooterDocuments.module.scss
@@ -1,0 +1,32 @@
+@import '@/theme/variables';
+@import '@/theme/mixins';
+
+.wrapper {
+  @media (max-width: $mobile) {
+    margin-bottom: 1.5rem;
+  }
+  ul {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+  }
+  a {
+    color: $white-text;
+    position: relative;
+  }
+  a:hover {
+    color: $background-two;
+  }
+  a::after {
+    content: '';
+    position: absolute;
+    left: 0;
+    bottom: -5px;
+    width: 100%;
+    height: 1px;
+    background-color: $white-text;
+  }
+  a:hover::after {
+    background-color: $white-text;
+  }
+}
